Fixed it..I need to look more into this
I had a variable list set as a shallow copy of a full slice [:] of
another list  and changes to the variable were affecting the first..This
was very strange because doing this with 1 dimensional lists never
caused this bug. I just wound up remaking the first list after doing the
variable set.